SPR/requirements.txt
pytest==8.4.1
pytest-cov==6.2.1
coverage==7.9.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.4.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
fpdf2==2.7.6
reportlab==4.0.7
pyttsx3==2.90
gTTS==2.4.0 